Use `nom_locate::LocatedSpan` as lexer input
[`nom_locate`](https://docs.rs/nom_locate/latest/nom_locate/) keeps
track of source file location of parsed input. This opens the
future possibility to include source file location in
error messages.

It could also be used to include source file location of parsed objects,
for example to include it in generated output comments or for more
details in later error messages.