Use a tri-state for InformationObjectClassField optionality
Information object class fields can be either OPTIONAL, DEFAULT or
required. Model it with an enum of these three states.

This makes the invalid state where `is_optional == true` and
`default` is `Some(..)` unrepresentable.

This also makes the default value generic, which will be required to
parse different types of default values in the future.