feat(lexer): More robust parsing of SymbolDefn in MACROs
Fix a problem where parsing for example `value(myValueRef MyType)` would
first match "myValueRef" as a `SymbolDefn::ValueMacroType`, and then
fail when next token is "MyType" instead of the expected ")".

Fixed by having the ending ")" be part of the opt's instead of using
`in_parentheses`.

Also changes so that `SymbolDefn::ValueLocalvaluereferenceMacroType` is
allowed to be a local_type_reference, in addition to
local_value_reference (for example allow `value(MyTypeRef MyType)`).
This is *not* according to the spec but it is used in the wild, and
should probably generate a warning somewhere (not implemented here).