Some improvements of lexer::utils::debug_result
If parsing was successful, print the resulting value, the matched
characters and the remaining input.

If parsing failed, print the error in human readable form.

Write the output to stderr so that it won't be mixed when using
`--stdout` output.

Use `nom::Parser` trait and be a bit more general.